共通レイアウトファイルを切り分け
